add client visitor test for related objects builder
Add missing placeholder attribute for SimpleInputComponent in ConstructVisitor.
